Safari-611.1.10/LayoutTests/http/tests/cookies/private-cookie-storage-expected.txt

This test checks that cookies set from a Set-Cookie header in private browsing mode are stored in the same cookie storage as cookies set with document.cookie.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Cookies after disabling private browsing (should be key1=value1):key1=value1
PASS cookie is 'key1=value1'.
PASS successfullyParsed is true

TEST COMPLETE
